Prevent crash when a target class is not located with better logging mechanism. (#29)
This PR fixes #27 by adding a new type of exception `TargetClassNotFound`. With this PR annotator will not crash if the class is not found, and will instead log this information for later debugging. 

Also added a test to verify, annotator will not fail and the desired error message is correctly reported in the `err` console.