Rename Scanner module to TypeAnnotatorScanner. (#36)
This PR resolves #25 and renames `Scanner` module to `AnnotatorScanner` module.

The name `Scanner` is too generic, specifically when mentioned along other `ErrorProneCheckers` and should be refined. Currently the suggestion is: `AnnotatorScanner`.


This involved changes in the configurations, unit tests project templates, and install scripts.

@lazaroclapp Previously you suggested `NullAwayAnnotatorScanner`, however, this checker is not related to NullAway and I am intending to apply the same technique and use this whole repo in general to infer other type systems. Given that this repo is called `NullAwayAnnotator`, I still think it makes sense to generalize this name. Please let me know your thoughts on this and if you think I should rename it to `NullAwayAnnotatorScanner`. Thanks :)

----
**Update**:
The final decision regarding name is: `TypeAnnotatorScanner`.