Improve computation of lower/upper bound of number of errors on downstream dependencies in fix trees (#46)
This PR resolves #45 by improving computation of the lower bound of number of errors on downstream dependencies in fix trees. 

With this change, annotator will consider the maximum of number of errors triggered by each node in the fix tree as the lower bound which. Previous approach only considers the effect of the root of tree which can be really far from the the real number. This PR improves that.

----
#### Update

In this PR an upper bound is also computed and serialized for experiements.