update unit test templates (#114)
This PR updates test infrastructure for `annotator-core` module to use the using gradle wrapper in the project itself. We currently use the gradle installed in the system itself to create an appropriate gradle wrapper for each test which is very inefficient. The PR updates the test infra to reuse the existing gradle in the project itself.

This PR updates all templates to use error prone `2.0.1` from `0.6` as the older version is not compatible with gradle version `7.4.2`.