Enhance error data structure. (#112)
This PR is build upon #103.

In NullAway, if the error is resolvable and involved a pseudo-assignment of a `@Nullable` expression into a `@NonNull` target, the location of that target is serialized. This information is enough to generate a fix resolving the reported error. Currently we only store the location of the `@Nonnull` element in `Error` data structure. This PR updates `Error` data structure to store the actual fix which can resolve the error.


Changes in this PR
- Update Error data structure to store the resolving fixes.
- Creates fixes according to the information serialized for the reporting error.