remove duplicate API (#122)
This PR removes `duplicate` method on `Location` instances which is unused in the source code and only used in one unit tests. 

This PR is straightforward and simply removes an used method, I am going to land this PR once CI passes, if there is anything you would like to discuss, please let me know and I will bring this back.

Please note, that this was required for #121 which resolves flaky CI issue.