Remove triggered fixes from iterations. (#113)
Following #112 errors contain the set of fixes which can resolve the reported errors. Currently, after each fix examination, the set of triggered fixes and errors are computed per region. Since #112 updates error instances to store the resolving fixes, the set of triggered fixes (which resolves the triggered errors) is redundant. This PR removes this structure.

Changes in this PR:

- Removes set of triggered fixes from `Nodes` in `ConflictGraph`.
- Constructs fix trees based on the containing resolving fixes of triggered errors.
- Adds set of triggered fixes from downstream to `nodes` for clear separation between fixes from downstream and fixes from target.