Update nullaway test version to 0.10.8 (#136)
This PR simply updates NullAway test version to `0.10.8` and fixes the revealed bug regarding initialization errors. One major update in `0.10.8` is serializing the actual constructor for initialization errors rather than just `null` which helped reveal this bug in tests.

This PR also update how we compute `origins` (initial set of set regions where a root fix is triggered). 


Previous to this PR we were not including initialization errors in a root fix origins, as they were reported in the null region, but with the new update on initialization serialization, we should include them (the condition is `singleFix` is removed).