Configure CI to run all tests in all configurations (#127)
This updates CI jobs to run all tests in all three configurations.

We have configurations below for running tests:
1. `Parallel Processing`: refers to computing effectiveness of multiple fixes simultaneously using graph coloring.
2. `Caching`: refers to caching fix’s impacts and reuse in next iterations.



This PR updates CI to run all tests in three combinations below with flags above:
|                                        |     Parallel Processing Disabled     |     Parallel Processing Enabled     |
| :---------------------: | :--------------------------------:| :-------------------------------: |
|  **Caching Disabled** |                           ✓                            |                           ✓                           | 
|  **Caching Enabled**  |                           X                             |                           ✓                           |  



In the current architecture of unit tests, we have a set of unit tests (that each will be converted to a standalone gradle project) which the internal state of Annotator is verified. Currently, both parallel processing and caching is activated. This PR updates the CI by rerunning unit tests to verify the internal state of the Annotator is consistent among these configurations as they are merely optimization and should not have any effect on the impact computation.