Update impacted region detection to include constructors for uninitialized fields. (#148)
This PR update the set of potentially impacted regions for uninitialized fields at declaration.

Following update to NullAway `0.10.8` initialization errors are reported in constructor regions rather than the `null` regions. For fields which are not initialized at the declaration, all constructors can be a potentially impacted region as if it is possible that the corresponding initialization error get resolved once the field is annotated as `@Nullable`.

Please consider example below:

```java
public class Foo {
    
     Object bar;
    
     public Foo() { } 
}
```

Since there is no trace of usage of `bar` in `Foo()`, currently we do not consider `Foo()` as a potentially impacted region for changes on `bar` whereas, if `bar` get annotated as `@Nullable`, an error will be resolved on `Foo()`. This PR updates the logic for detecting potentially impacted region for uninitialized fields at declaration. 


Please note that for initialized fields at declaration, constructors are not responsible for initializing the field, therefore, we do not get any initialization errors on constructors.

This PR updates the internal data structure for `FieldDeclarationStore` to keep a record of initialized fields at declaration and `MethodDeclarationTree` to identify constructors.