Add 1 unit tests for nested parameter case (#138)
This PR is a followup for #111 which only adds two 1 test to ensure correct set of regions are selected as potentially impacted regions for a change on a parameter. For parameters, the call sites of the method, can be a potentially impacted region as well.

Consider scenario below:

```java
void foo(Object p){
       bar(p);
}

void bar(Object p)
```

assume that we have two fixes on foo#bar and `bar#p` and `foo#p`, while we are analyzing impact of making `foo#p` nullable, `bar#p` should not be annotated as nullable, otherwise we will miss the triggered error that p is nullable while invoking `bar`. This addition in effective region will make a region conflict for `foo#p` and `bar#p` and they will not be processed simultaneously.