Update NullAway test version to 0.10.10 (#149)
This PR updates NullAway test version to `0.10.10`. The main update in this version is on the serialization change. 

In the newest version, method signatures does not contain any annotation or type arguments, this PR adapts injector to work with this serialization to locate the method and also update scanner to use the same logic for serializing method signatures.


Please note that previously Injector was written to parse complicated cases such as case below:

```
foo(@edu.ucr.Custom({"arg1, "arg2}) java.util.Map<A.B.C, A.B.D>)
```

Wheres in the latest version of serialization for the method above, we receive input below:
```
foo(java.util.Map)
```

Therefore, the parsing logic, along the input for some unit tests had to be updated.