Add mocked build to unit tests (#178)
This PR prepares unit tests for changes in #175 by adding a mocked build to some unit tests which requires outputs of `Scanner` serialization just to finalize the `Config` instance. The content are not important but the files should only exist after a build request during the construction of the `config` instance. This PR update some unit tests to use the mocked build procedure to make the desired outputs.