Update deserializer interface (#181)
This PR updates the `Adapter` interface for deserialization of errors and outputs from Annotator Scanner. It also renames it to `CheckerDeserializer` which is preparation for the upcoming changes toward adapting Annotator to work with `UCRTaint` analysis. 

Annotator Scanner outputs does not need to be deserialized through adapters, as we can safely assume that the matching version of Annotator and its Scanner is always used within in the execution, therefore it is moved simply to `Utility` class.