Rename Tracker to RegionRegistry (#195)
`Trackers` are used to retrieve potentially impacted regions by changes in a fix tree. However, the content of these data structures are created only once. The name tracker for this data structures does not fully reflect their usage, therefore, in this PR they are changed to `RegionRegistry` as it better reflects that their content are immutable and created at the beginning of execution.