Update Library Model Loader to mark fields Nullable (#220)
Followup on an update in NullAway recent change [#878](https://github.com/uber/NullAway/pull/878) which enables NullAway to mark fields as `@Nullable`. This PR updates our `LibraryModelsLoader` to enable communication of Annotator and NullAway to mark fields `@Nullable` while analyzing downstream dependencies. This is required to prepare the code for the upcoming change which enables Annotator to be informed of impacts of making fields `@Nullable` on downstream dependencies.