Switch to Spotless for formatting (#249)
We were using another plugin for formatting before, but Spotless is maintained.  Also we update to latest GJF version, which tweaks the formatting a bit.

No need to merge this immediately if it causes lots of conflicts in outstanding PRs.  We can always do it later and re-run the formatting pass.

Fixes #248