Refines type argument annotation operations (#252)
Previous to this PR, if a type-use annotation was requested to add by the Injector, it would add the annotation on all the type arguments and add the declaration recursively. With this PR, we can control exactly determine the location of arguments where should be annotated. It uses the index system to locate the target type arg.