Insert annotations at the correct position for arrays (#251)
This PR helps insert annotations at the correct position for array declarations.

For array declarations we want the annotations to be inserted just before the array type brackets. For example : 
String `@Nullable [] f2;`

Currently, the behaviour is that the annotations are inserted before the simpleName type. Like this :
`@Nullable String[] f2;`