fix array component primitive type (#257)
This PR fixes a bug in handling type arguments where arrays with primitive component types were incorrectly processed. The issue stemmed from an assumption that type arguments could not be primitives. However, we missed the case where an array used as a type argument can have a primitive component type.

To resolve this, the relevant method in TypeArgChangeVisitor is now overridden to account for arrays with primitive component types used as type arguments. This ensures that the appropriate changes are applied consistently.